Fix race condition
If two processes were instantiated to the same repository and received
capabilities roughly at the same time, causing one to wait for the other's pull
to finish, the subsequent pull would be performed on a "dirty" instance of the
repository, possibly causing repository corruption. This is fixed by taking a
lock at the outer level, invalidating, then performing the pull with the lock
already in place.